Rename "rooms_attributes" to "rooms_gbc_overlays", reformat
"rooms_attributes" is actually a byte-per-tile array of actual
tiles to use on the overworld for each screen. It is used in GBC
mode to determine what tile is actually shown, regardless of the
actual physical tile underneath.

Used in some places e.g. Mabe Village to give the fancy colored
patches of bordered ground (which take up the upper object values,
used for tree macros and the like in actual layout objects)