wram: separate wOAMBuffer and wLinkOAMBuffer again
The OAMBuffer area can be used for several purposes.

One is on the World gameplay mode, where indeed part of it is allocated
for Link sprites, and part of it for entities.

But it is also used in other gameplay modes, like cutscenes, menus, etc
to store a variety of arbitrary sprites.

So we need to keep both variables names (wOAMBuffer and wLinkOAMBuffer),
even if they refer to the same address technically. In the code,
sometimes the code purpose is clearly to update Links sprites (and in
this case wLinkOAMBuffer is better used). But otherwise, when the code
is simply trying to manipulate arbitrary OAM data, wOAMBuffer is a
better fit.