tools: add --no-file-terminator to convert_background.py
Partial tilemaps don't end with a NULL terminator. Instead, commands
cascade into the next tilemap stored in ROM or RAM.

This is used mostly by menu commands, which overwrite the larger menu
tilemaps.

When decoding, the `--no-file-terminator` option tells the decoder that
the file is expected to end abruptely.

And when encoding, the `--no-file-terminator` option tells the encoder
not to append the NULL file terminator at the end of the file.

Fix #448